Move images into assets folder
