+++ tsconfig.json
@@ -10,5 +10,6 @@
     "strict": true,
     "skipLibCheck": true
   },
-  "files": ["./index.ts"]
+  "files": ["./index.ts"],
+  "extends": "expo/tsconfig.base"
 }
